+++ .docker/dime
@@ -1,0 +1,29 @@
+#!/bin/bash
+
+if ! command -v docker >/dev/null 2>&1; then
+    echo
+    echo "Please check that Docker is correctly installed https://www.docker.com"
+    echo
+    exit
+fi
+
+if [[ "$1" == "delete" ]]; then
+    docker rmi -f ghcr.io/osteel/dime
+    rm -- "$0"
+    exit
+fi
+
+if ! [ -f "~/.dime/database.sqlite" ]; then
+    mkdir -p ~/.dime && touch ~/.dime/database.sqlite
+fi
+
+command="docker run -it --rm -v ~/.dime/database.sqlite:/root/.dime/database.sqlite"
+
+if [[ "$1" == "process" && -n "$2" && -f "$2" ]]; then
+    filename=$(basename "$2")
+    command+=" -v $2:/tmp/$filename ghcr.io/osteel/dime:latest process /tmp/$filename"
+else
+    command+=" ghcr.io/osteel/dime:latest ${@:1}"
+fi
+
+eval "$command"
