+++ .github/workflows/release.yml
@@ -2,15 +2,18 @@ name: Release
 
 on:
   push:
-    tags: ["*"]
+    tags: ["v*"]
+  pull_request:
+    branches: [main]
 
 env:
   APP_ENV: "local"
+  REGISTRY: "ghcr.io"
 
 jobs:
 
-  build:
-    name: Build
+  phar:
+    name: Phar
     runs-on: ubuntu-latest
 
     steps:
@@ -31,7 +34,8 @@ jobs:
         uses: ramsey/composer-install@v2
 
       - name: Build archive
-        run: "php dime app:build dime --build-version=${{ github.ref_name }}"
+        run: "php dime app:build dime --build-version=v0.1.2"
+        #run: "php dime app:build dime --build-version=${{ github.ref_name }}"
 
       - name: Draft release
         uses: softprops/action-gh-release@v1
@@ -41,10 +45,41 @@ jobs:
           generate_release_notes: true
 
       - name: Update README
-        run: "php release.php ${{ github.ref_name }}"
+        run: "php .github/release.php v0.1.2"
+        #run: "php .github/release.php ${{ github.ref_name }}"
 
       - name: Push changes
         uses: stefanzweifel/git-auto-commit-action@v4
         with:
-          branch: main
           commit_message: "Updated README"
+
+  docker:
+    name: Docker
+    needs: phar
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Log into registry
+        uses: docker/login-action@v2
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.repository_owner }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Set up Buildx
+        uses: docker/setup-buildx-action@v2
+
+      - name: Build and push
+        uses: docker/build-push-action@v4
+        with:
+          context: .
+          file: .docker/Dockerfile
+          push: true
+          platforms: linux/amd64,linux/arm64
+          tags: ${{ env.REGISTRY }}/${{ github.repository }}:latest
