--- app/Services/SelfUpdate/Helpers.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Services\SelfUpdate;
-
-final class Helpers
-{
-    public static function installedViaComposer(): bool
-    {
-        return str_contains(__DIR__, '.composer/vendor');
-    }
-}
